requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
beautifulsoup4==4.14.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cramjam==2.11.0
cssselect==1.3.0
curl_cffi==0.13.0
dotenv==0.9.9
fastparquet==2024.11.0
finviz==1.4.6
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2025.9.0
idna==3.10
iniconfig==2.1.0
lxml==6.0.2
multidict==6.6.4
multitasking==0.0.12
numpy==2.3.3
packaging==25.0
pandas==2.3.3
peewee==3.18.2
platformdirs==4.4.0
pluggy==1.6.0
propcache==0.3.2
protobuf==6.32.1
pyarrow==21.0.0
pycparser==2.23
Pygments==2.19.2
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
six==1.17.0
soupsieve==2.8
tenacity==9.1.2
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
user_agent==0.1.14
websockets==15.0.1
yarl==1.20.1
yfinance==0.2.66
